- fixed bug with dropdown not being deleted from editor; - fixed bug with text not being inserted when dropdown is closed; - small refactoring of CustomPortal -> PortalAtCursorPosition
